requirements.txt
pandas
openpyxl
matplotlib
pyserial
pillow
tkinter